+++ libraries/test_utils/src/main/java/androidx/media3/test/utils/FakeRenderer.java
@@ -61,7 +61,9 @@ public class FakeRenderer extends BaseRenderer {
   private boolean hasPendingBuffer;
   private List<Format> formatsRead;
 
+  public boolean isInitialized;
   public boolean isEnded;
+  public boolean isReleased;
   public int positionResetCount;
   public int sampleBufferReadCount;
   public int enabledCount;
@@ -195,5 +197,15 @@ public class FakeRenderer extends BaseRenderer {
    */
   protected boolean shouldProcessBuffer(long bufferTimeUs, long playbackPositionUs) {
     return bufferTimeUs < playbackPositionUs + SOURCE_READAHEAD_US;
+  }
+
+  @Override
+  protected void onInit() {
+    isInitialized = true;
+  }
+
+  @Override
+  protected void onRelease() {
+    isReleased = true;
   }
 }